requirements.txt
streamlit
langchain
langchain-community
qdrant-client
pdf2image
pytesseract
PyMuPDF
python-dotenv
